refactor from_db_notification into its own encoder module
